Moved lookup table packer out of detail namespace. (#235)
* Moved lookup_table_definition.hpp

* Moved lookup table packer out of detail:: namespace.

* Commented the printing out.
